Created dedicated file to hold functions for tests
