+++ uvdata/tests/test_uvdata.py
@@ -12,61 +12,7 @@ import warnings
 import collections
 from astropy.io import fits
 import sys
-
-
-def get_iterable(x):
-    if isinstance(x, collections.Iterable):
-        return x
-    else:
-        return (x,)
-
-
-def clearWarnings():
-    # Quick code to make warnings reproducible
-    for name, mod in list(sys.modules.items()):
-        try:
-            reg = getattr(mod, "__warningregistry__", None)
-        except ImportError:
-            continue
-        if reg:
-            reg.clear()
-
-
-def checkWarnings(obj, func, func_args=[], func_kwargs={},
-                  warning_cat=UserWarning,
-                  nwarnings=1, warning_message=None, known_warning=None):
-
-    if known_warning == 'miriad':
-        # The default warnings for known telescopes when reading miriad files
-        warning_cat = [UserWarning]
-        warning_message = ['Altitude is not present in Miriad file, using known '
-                           'location values for PAPER.']
-        nwarnings = 1
-    elif known_warning == 'paper_uvfits':
-        # The default warnings for known telescopes when reading uvfits files
-        warning_cat = [UserWarning] * 2
-        warning_message = ['Required Antenna frame keyword', 'telescope_location is not set']
-        nwarnings = 2
-
-    warning_cat = get_iterable(warning_cat)
-    warning_message = get_iterable(warning_message)
-
-    clearWarnings()
-    with warnings.catch_warnings(record=True) as w:
-        warnings.simplefilter("always")  # All warnings triggered
-        status = func(*func_args, **func_kwargs)  # Run function
-        # Verify
-        if len(w) != nwarnings:
-            print('wrong number of warnings')
-            for idx, wi in enumerate(w):
-                print('warning {i} is: {w}'.format(i=idx, w=wi))
-            obj.assertTrue(False)  # Fail the test, move on
-        else:
-            for i, w_i in enumerate(w):
-                obj.assertIs(w_i.category, warning_cat[i])
-                if warning_message[i] is not None:
-                    obj.assertIn(warning_message[i], str(w_i.message))
-    return status
+from test_functions import *
 
 
 test_file_directory = '../data/test/'
